pyproject.toml
# Copyright 2025 Poke & Wiggle GmbH. All rights reserved.
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     https://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

[project]
name = "example_policies"
version = "0.1.0"
authors = [
    {name = "Maximilian Li", email="max@pokeandwiggle.com"}
]
maintainers = [
    {name = "Poke & Wiggle GmbH", email="info@pokeandwiggle.com"}
]
readme = "README.md"
dependencies = [
    "grpcio>=1.74.0",
    "grpcio-tools>=1.74.0",
    "lerobot[smolvla]",
    "matplotlib>=3.10.6",
    "mcap>=1.3.0",
    "opencv-python>=4.12.0.88",
    "rosbags>=0.10.11",
    "scipy>=1.15.3",
    "torch>=2.8.0",
    "torchcodec>=0.6.0",
    "torchvision>=0.23.0",
]

[tool.uv.sources]
lerobot = { git = "https://github.com/LiXiling/lerobot.git" }


[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
